fitness_agent/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
google-generativeai
pinecone
sentence-transformers